Add profile after login
commmit includes add profile page to direct to after login. This allows users to set up their profile page or skip it for now. Also added edit page, so that once on their profile they can choose to edit it. Nothing is functioning but most of the grunt work should be done.