issue-42 #2
I tried implementing a way to derive a URL from a path using windows.location.href but learned that it is only for the current web page. Slack does not have a web page so I might have to just write in the link once we have the website up.

Notice: Write in link once Help Page is up. Talk to Moore about this.